feat: onboarding for frontend operator in consoledot
These changes will support the move to the frontend operator for this
app.

- package.json edited to disable the jest cache in test
- pr_check.sh setup for verifying the app can build
- build_deploy.sh for building and committing the image to the repo

Once this is all wired in, we should be able to remove the midstream
step that is required to manage releases. Once onboarded, we can use the
git ref from github to place images in prod or stage.

Signed-off-by: Stephen Adams <stephen.adams@redhat.com>